fully working for first two mazes with camera and cartographer

+++ turtlebot3_gazebo/include/color_detector.hpp
@@ -5,7 +5,7 @@
 #include <sensor_msgs/msg/image.hpp>
 #include <cv_bridge/cv_bridge.h>
 #include <opencv2/opencv.hpp>
-#include "std_msgs/msg/float32.hpp"
+#include "std_msgs/msg/float64.hpp"
 #include "std_msgs/msg/bool.hpp"
 
 class ColorDetector : public rclcpp::Node
@@ -17,7 +17,7 @@ private:
     void image_callback(sensor_msgs::msg::Image::SharedPtr msg);
 
     // Add the publishers for green percentage and detection status
-    rclcpp::Publisher<std_msgs::msg::Float32>::SharedPtr green_percentage_pub_;
+    rclcpp::Publisher<std_msgs::msg::Float64>::SharedPtr green_percentage_pub_;
     rclcpp::Publisher<std_msgs::msg::Bool>::SharedPtr green_detected_pub_;
 
     rclcpp::Subscription<sensor_msgs::msg::Image>::SharedPtr subscription_;
